refactor: 패키지 구조 변경 및 ChessBoardGenerator utils 패키지 변경
