--- src/main/java/domain/chessboard/BoardGenerator.java
@@ -1,10 +1,0 @@
-package domain.chessboard;
-
-import domain.piece.Piece;
-import domain.position.Position;
-import java.util.Map;
-
-public interface BoardGenerator {
-
-    Map<Position, Piece> generate();
-}
