+++ src/main/java/domain/piece/King.java
@@ -1,7 +1,7 @@
 package domain.piece;
 
 import domain.Player;
-import domain.directions.Direction;
+import domain.direction.Direction;
 import java.util.List;
 
 public class King extends SpecificLocationPiece {
