+++ src/main/java/domain/piece/SpecificLocationPiece.java
@@ -1,7 +1,7 @@
 package domain.piece;
 
 import domain.Player;
-import domain.directions.Direction;
+import domain.direction.Direction;
 import domain.position.Position;
 import java.util.ArrayList;
 import java.util.List;
